Added support for expr in autox + dnd in pexplorer is in progress

+++ yukthi-autox/src/main/java/com/yukthitech/autox/AutomationContext.java
@@ -174,7 +174,11 @@ public class AutomationContext
 		
 		try
 		{
-			AutomationUtils.deleteFolder(workDirectory);
+			if(workDirectory.exists())
+			{
+				AutomationUtils.deleteFolder(workDirectory);
+			}
+			
 			FileUtils.forceMkdir(workDirectory);
 		}catch(Exception ex)
 		{
